+++ Tarlad/Models/User.swift
@@ -16,21 +16,21 @@ public class User: NSManagedObject {
         return NSFetchRequest<User>(entityName: "User")
     }
     
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Int
     @NSManaged public var nickname: String
     @NSManaged public var name: String
     @NSManaged public var surname: String
     @NSManaged public var imageURL: String?
     
     public func setData(item: [String: Any]) {
-        id = item["id"] as! Int64
+        id = item["id"] as! Int
         nickname = item["nickname"] as! String
         name = item["name"] as! String
         surname = item["surname"] as! String
         imageURL = item["imageURL"] as? String
     }
     
-    public class func getId(item: [String: Any]) -> Int64 {
-        return item["id"] as! Int64
+    public class func getId(item: [String: Any]) -> Int {
+        return item["id"] as! Int
     }
 }
